requirements.txt
efficientnet==1.1.1
colorama==0.4.3
jupyter==1.0.0
keras==2.2.5
numpy==1.16.2
opencv-python==4.4.0.42
pandas==0.25.3
Pillow==7.2.0
scikit-learn
scikit-image
scikit-plot
scipy
tensorflow-gpu==1.14.0

